+++ lib/components/screens/infoscreen.dart
@@ -34,7 +34,7 @@ class InfoScreen extends Component
     TextComponent title = TextComponent(
         text: "Eile mit Weile", textRenderer: textPaint, anchor: Anchor.center);
 
-    title.position = Vector2(game.size.x / 2, scale.y * 150);
+    title.position = Vector2(game.size.x / 2, scale.y * 250);
     title.scale = scale;
 
     add(title);
@@ -49,7 +49,7 @@ class InfoScreen extends Component
         textRenderer: textPaint_small,
         anchor: Anchor.center);
 
-    subtitle.position = Vector2(game.size.x / 2, scale.y * 300);
+    subtitle.position = Vector2(game.size.x / 2, scale.y * 400);
     subtitle.scale = scale;
     add(subtitle);
 
@@ -69,7 +69,7 @@ class InfoScreen extends Component
         size: Vector2(game.size.x - 100, game.size.y),
         anchor: Anchor.topCenter);
 
-    help.position = Vector2(game.size.x / 2, scale.y * 400);
+    help.position = Vector2(game.size.x / 2, scale.y * 500);
     help.scale = scale;
 
     add(help);