+++ lib/components/screens/maingame.dart
@@ -1,4 +1,3 @@
-import 'package:EileMitWeile/components/box.dart';
 import 'package:EileMitWeile/components/sprites/home_field.dart';
 import 'package:EileMitWeile/components/sprites/move_buttons.dart';
 import 'package:EileMitWeile/components/token.dart';
@@ -6,35 +5,18 @@ import 'package:EileMitWeile/components/gamecreation.dart';
 import 'package:flame/components.dart';
 
 import 'package:flame/experimental.dart';
-import 'package:flame/palette.dart';
-
-import 'package:flutter/rendering.dart';
 
 import '../../eilemitweile_game.dart';
 import '../../enums.dart';
 import '../box_dice.dart';
 import '../field.dart';
 import '../player.dart';
-import '../text_components/kills.dart';
 import '../text_components/state.dart';
 
 class MainGame extends Component
     with TapCallbacks, HasGameRef<EileMitWeileGame> {
   @override
   Future<void> onLoad() async {
-    int humans = 0;
-    for (Player player in gameRef.players) {
-      if (!player.is_AI) {
-        humans++;
-      }
-    }
-
-    if (humans > 1) {
-      gameRef.is_hotseat = true;
-    } else {
-      gameRef.is_hotseat = false;
-    }
-
     Sprite movebuttonSprite1 = emwSprite(1150, 0, 332, 332);
 
     for (var i = 0; i < 4; i++) {
@@ -51,24 +33,7 @@ class MainGame extends Component
       gameRef.move_buttons.add(button);
     }
 
-    if (gameRef.is_hotseat == true) {
-      for (var i = 0; i < 4; i++) {
-        ButtonComponent button = ButtonComponent(
-          position: Vector2(EileMitWeileGame.info_col_size,
-              EileMitWeileGame.fieldHeight + i * 350),
-          sprite: movebuttonSprite1,
-        );
-
-        button.button_number = i;
-
-        //button.size = Vector2(350, 350);
-        gameRef.move_buttons.add(button);
-      }
-    }
-
     gameRef.fields = CreateFields();
-
-    //gameRef.players = CreatePlayers();
 
     final List<HomeField> home_fields = [];
 
@@ -206,17 +171,17 @@ class MainGame extends Component
       }
     }
 
-    Box box = Box();
-
     InfoButton infoButton = InfoButton(
         position: Vector2(
-            EileMitWeileGame.info_col_size + EileMitWeileGame.fieldHeight,
-            gameRef.screenHeight - 140));
+            EileMitWeileGame.info_col_size, gameRef.screenHeight - 140));
 
     AudioButton audioButton = AudioButton(
         position: Vector2(
-            EileMitWeileGame.info_col_size + EileMitWeileGame.fieldHeight + 130,
-            gameRef.screenHeight - 140));
+            EileMitWeileGame.info_col_size + 130, gameRef.screenHeight - 140));
+
+    MenuButton menuButton = MenuButton(
+        position: Vector2(
+            EileMitWeileGame.info_col_size + 260, gameRef.screenHeight - 140));
 
     gameRef.heaven.size = Vector2(600, 600);
 
@@ -234,86 +199,27 @@ class MainGame extends Component
             20,
         gameRef.screenHeight / 2);
 
-    gameRef.world.add(box);
+    gameRef.world.add(gameRef.box);
     //gameRef.world.add(version);
-    box.addAll(gameRef.players);
-    box.addAll(gameRef.fields);
+    gameRef.box.addAll(gameRef.players);
+    gameRef.box.addAll(gameRef.fields);
 
-    box.add(gameRef.state_text_left = State.stateInfo());
-    box.add(gameRef.state_text_right = State.stateInfo());
-    box.add(box_dice_left);
-    box.add(box_dice_right);
+    gameRef.box.add(gameRef.state_text_left = State.stateInfo());
+    gameRef.box.add(gameRef.state_text_right = State.stateInfo());
+    gameRef.box.add(box_dice_left);
+    gameRef.box.add(box_dice_right);
 
-    box.addAll(home_fields);
-    box.addAll(gameRef.move_buttons);
-    box.add(gameRef.heaven);
-    box.addAll(heaven_fields0);
-    box.addAll(heaven_fields1);
-    box.addAll(heaven_fields2);
-    box.addAll(heaven_fields3);
-    box.addAll(tokens);
-    box.add(infoButton);
-    box.add(audioButton);
-
-    if (game.is_hotseat) {
-      box.add(gameRef.kill_text = KillInfo.killInfo());
-      gameRef.kill_text.position =
-          Vector2(EileMitWeileGame.console, gameRef.screenHeight - 100);
-    } else {
-      Sprite info_sprite = emwSprite(628, 841, 330, 330);
-
-      SpriteComponent info_box =
-          SpriteComponent(sprite: info_sprite, anchor: Anchor.topLeft);
-      info_box.size = Vector2(400, 400);
-      info_box.position = Vector2(EileMitWeileGame.info_col_size - 25,
-          900 + 3 * EileMitWeileGame.fieldHeight + 25);
-      box.add(info_box);
-
-      final style_info = TextStyle(
-          color: BasicPalette.black.color,
-          fontSize: 35,
-          fontFamily: 'PolandFull');
-
-      TextPaint textPaint_info = TextPaint(style: style_info);
-
-      TextComponent round_text = TextComponent(
-          text: "Rnd", textRenderer: textPaint_info, anchor: Anchor.center);
-      round_text.position = Vector2(EileMitWeileGame.info_col_size + 80,
-          900 + 4 * EileMitWeileGame.fieldHeight + 55);
-      box.add(round_text);
-
-      gameRef.round_num = TextComponent(
-          text: "0", textRenderer: textPaint_info, anchor: Anchor.center);
-      gameRef.round_num.position = Vector2(EileMitWeileGame.info_col_size + 245,
-          900 + 4 * EileMitWeileGame.fieldHeight + 55);
-      box.add(gameRef.round_num);
-
-      gameRef.kill_red = TextComponent(
-          text: "0", textRenderer: textPaint_info, anchor: Anchor.center);
-      gameRef.kill_red.position = Vector2(EileMitWeileGame.info_col_size + 120,
-          900 + 4 * EileMitWeileGame.fieldHeight + 190);
-      box.add(gameRef.kill_red);
-
-      gameRef.kill_blue = TextComponent(
-          text: "0", textRenderer: textPaint_info, anchor: Anchor.center);
-      gameRef.kill_blue.position = Vector2(EileMitWeileGame.info_col_size + 120,
-          900 + 4 * EileMitWeileGame.fieldHeight + 290);
-      box.add(gameRef.kill_blue);
-
-      gameRef.kill_green = TextComponent(
-          text: "0", textRenderer: textPaint_info, anchor: Anchor.center);
-      gameRef.kill_green.position = Vector2(
-          EileMitWeileGame.info_col_size + 285,
-          900 + 4 * EileMitWeileGame.fieldHeight + 290);
-      box.add(gameRef.kill_green);
-
-      gameRef.kill_purple = TextComponent(
-          text: "0", textRenderer: textPaint_info, anchor: Anchor.center);
-      gameRef.kill_purple.position = Vector2(
-          EileMitWeileGame.info_col_size + 285,
-          900 + 4 * EileMitWeileGame.fieldHeight + 190);
-      box.add(gameRef.kill_purple);
-    }
+    gameRef.box.addAll(home_fields);
+    gameRef.box.addAll(gameRef.move_buttons);
+    gameRef.box.add(gameRef.heaven);
+    gameRef.box.addAll(heaven_fields0);
+    gameRef.box.addAll(heaven_fields1);
+    gameRef.box.addAll(heaven_fields2);
+    gameRef.box.addAll(heaven_fields3);
+    gameRef.box.addAll(tokens);
+    gameRef.box.add(infoButton);
+    gameRef.box.add(audioButton);
+    gameRef.box.add(menuButton);
 
     gameRef.state_text_left.position = Vector2(
         EileMitWeileGame.info_col_size / 2 - 20, gameRef.screenHeight / 2);
@@ -327,20 +233,7 @@ class MainGame extends Component
             20,
         gameRef.screenHeight / 2);
 
-    gameRef.move_buttons[1].setPlayerColor(1);
-    gameRef.move_buttons[2].setPlayerColor(2);
-    gameRef.move_buttons[3].setPlayerColor(3);
-    if (gameRef.is_hotseat == true) {
-      gameRef.move_buttons[5].setPlayerColor(1);
-      gameRef.move_buttons[6].setPlayerColor(2);
-      gameRef.move_buttons[7].setPlayerColor(3);
-    }
     await add(gameRef.world);
-
-    gameRef.current_player = gameRef.players[3];
-    //gameRef.players[0].is_AI = false;
-
-    gameRef.NextPlayer();
 
     final camera = CameraComponent(world: gameRef.world)
       ..viewfinder.visibleGameSize =
@@ -355,7 +248,7 @@ class InfoButton extends SpriteComponent
     with TapCallbacks, HasGameRef<EileMitWeileGame> {
   InfoButton({
     required Vector2 position,
-  }) : super(position: position, size: Vector2(100, 100)) {}
+  }) : super(position: position, size: Vector2(85, 85)) {}
 
   @override
   Future<void> onLoad() async {
@@ -373,7 +266,7 @@ class AudioButton extends SpriteComponent
     with TapCallbacks, HasGameRef<EileMitWeileGame> {
   AudioButton({
     required Vector2 position,
-  }) : super(position: position, size: Vector2(100, 100)) {}
+  }) : super(position: position, size: Vector2(85, 85)) {}
 
   @override
   Future<void> onLoad() async {
@@ -390,5 +283,23 @@ class AudioButton extends SpriteComponent
       gameRef.audio_enabled = true;
       this.sprite = emwSprite(724, 1183, 70, 70);
     }
+  }
+}
+
+class MenuButton extends SpriteComponent
+    with TapCallbacks, HasGameRef<EileMitWeileGame> {
+  MenuButton({
+    required Vector2 position,
+  }) : super(position: position, size: Vector2(85, 85)) {}
+
+  @override
+  Future<void> onLoad() async {
+    Sprite the_sprite = emwSprite(881, 1183, 70, 70);
+    this.sprite = the_sprite;
+  }
+
+  @override
+  void onTapUp(TapUpEvent event) {
+    gameRef.router.pushNamed('newgame');
   }
 }